Added dl2 to dl3 method corrected (in principle)

+++ data_processing_srunwise/script_dl1_to_dl3.py
@@ -9,9 +9,9 @@ from scipy.stats import chi2
 from scipy import optimize
 import subprocess
 
-from astropy.coordinates     import SkyCoord
-from lstchain.io.config      import get_standard_config
-from ctapipe.io              import read_table
+from astropy.coordinates import SkyCoord
+from lstchain.io.config  import get_standard_config
+from ctapipe.io          import read_table
 import tables
 
 # Other auxiliar scripts
@@ -32,8 +32,10 @@ source_name = "crab"
 
 # Standard paths for data in the IT cluster ---------
 root_dl1 = "/fefs/aswg/data/real/DL1/*/v0.*/tailcut84/"
-root_rfs = "/fefs/aswg/data/models/AllSky/20230901_v0.10.4_allsky_base_prod/"
-root_mcs = "/fefs/aswg/data/mc/DL2/AllSky/20230901_v0.10.4_allsky_base_prod/TestingDataset/"
+# root_rfs = "/fefs/aswg/data/models/AllSky/20240131_allsky_v0.10.5_all_dec_base/"
+root_rfs = "/fefs/aswg/data/models/AllSky/20230927_v0.10.4_crab_tuned/"
+# root_mcs = "/fefs/aswg/data/mc/DL2/AllSky/20240131_allsky_v0.10.5_all_dec_base/TestingDataset/"
+root_mcs = "/fefs/aswg/data/mc/DL2/AllSky/20230927_v0.10.4_crab_tuned/TestingDataset/"
 
 # Root path of this script
 root = os.getcwd() + "/"
@@ -56,62 +58,30 @@ dir_irfs        = root_data + "irfs/"
 def configure_lstchain():
     """Creates a file of standard configuration for the lstchain analysis. 
     It can be changed inside this function"""
-    dict_config = {
-      "EventSelector": {
-        "filters": {
-          "intensity": [80, "Infinity"],
-          "width": [0, "Infinity"],
-          "length": [0, "Infinity"],
-          "r": [0, 1],
-          "wl": [0, 1],
-          "leakage_intensity_width_2": [0, 1],
-          "event_type": [32, 32]
-        }
-      },
-      "DL3Cuts": {   
-        "min_event_p_en_bin": 100,
-        "min_gh_cut": 0.1,
-        "max_gh_cut": 0.95,
-        "min_theta_cut": 0.05,
-        "max_theta_cut": 0.32,
-        "fill_theta_cut": 0.32,
-        "allowed_tels": [1]
-      },
-      "DataBinning": {
-        "true_energy_min": 0.002,
-        "true_energy_max": 200,
-        "true_energy_n_bins_per_decade": 5,
-        "reco_energy_min": 0.002,
-        "reco_energy_max": 200,
-        "reco_energy_n_bins_per_decade": 5,
-        "energy_migration_min": 0.2,
-        "energy_migration_max": 5,
-        "energy_migration_n_bins": 31,
-        "fov_offset_min": 0.1,
-        "fov_offset_max": 1.1,
-        "fov_offset_n_edges": 9,
-        "bkg_fov_offset_min": 0,
-        "bkg_fov_offset_max": 10,
-        "bkg_fov_offset_n_edges": 21,
-        "source_offset_min": 0.0001,
-        "source_offset_max": 1.0001,
-        "source_offset_n_edges": 1000
-      }
-    }
+    dict_config = get_standard_config()
+    # We select the heuristic flatfield option in the standard configuration
+    dict_config["source_config"]["LSTEventSource"]["use_flatfield_heuristic"] = True
     with open(config_file, "w") as json_file:
         json.dump(dict_config, json_file)
 
-def main(input_str):
+def main(input_str, flag_scaled_str):
 
     # Run numbers to analyze, in our case only one
     run_number = int(input_str)
+
+    # Reading the scaled or not flag
+    if flag_scaled_str == "True":
+        flag_scaled = True
+    elif flag_scaled_str == "False":
+        flag_scaled = False
+    else:
+        logger.error(f"Input string for scaling: {flag_scaled_str} not valid.\nInput 'True' or 'False'")
     
     # Number of subruns to analyze per run
     subruns_num = None  # Specify the number of subruns you want to analyze, set subruns_num = None to analyze all subruns
 
-
     dir_dl3_scaled = dir_dl3_scaled_base + f"{run_number:05}/"
-    dir_dl3        = dir_dl3_base + f"{run_number:05}/"
+    dir_dl3        = dir_dl3_base        + f"{run_number:05}/"
 
     # Creating the directories in case they don't exist
     for path in [os.path.dirname(config_file), dir_dl1b_scaled, dir_dl1m_scaled, dir_dl2, dir_dl2_scaled, dir_dl3_scaled, dir_dl3, dir_irfs]:
@@ -203,68 +173,68 @@ def main(input_str):
 
 
 
-
-    for ir, run in enumerate(dict_dchecks.keys()):
-    
-        dir_run = dir_dl1b_scaled + f"{run:05}" + "/"
-        output_fname = dir_dl1m_scaled + f"dl1_LST-1.Run{run:05}.h5"
-    
-        command = f"lstchain_merge_hdf5_files --input-dir {dir_run} --output-file {output_fname} --run-number {run} --no-image"
-        logger.info(command)
+    if flag_scaled:
+        for ir, run in enumerate(dict_dchecks.keys()):
         
-        subprocess.run(command, shell=True)
+            dir_run = dir_dl1b_scaled + f"{run:05}" + "/"
+            output_fname = dir_dl1m_scaled + f"dl1_LST-1.Run{run:05}.h5"
         
-        dict_dchecks[run]["dl1b_scaled"]["runwise"] = output_fname
+            command = f"lstchain_merge_hdf5_files --input-dir {dir_run} --output-file {output_fname} --run-number {run} --no-image"
+            logger.info(command)
+            
+            subprocess.run(command, shell=True)
+            
+            dict_dchecks[run]["dl1b_scaled"]["runwise"] = output_fname
 
 
 
 
 
-
-    for ir, run in enumerate(dict_dchecks.keys()):
-    
-        input_fname  = dict_dchecks[run]["dl1b_scaled"]["runwise"]
-        output_fname = dir_dl2_scaled + input_fname.split("/")[-1].replace("dl1", "dl2", 1)
-        rf_node      = dict_dchecks[run]["simulations"]["rf"]
-    
-        # Check if the file exists and delete if exists (may be empty or half filled)
-        if os.path.exists(output_fname):
-            logger.info(f"File already exists, deleting and re-computing:\n-->{output_fname}")
-            os.remove(output_fname)
-    
-        logger.info(f"\nComputing dl2 for Run {run:5} (scaled data)")
-        logger.info(f"--> {output_fname}\n")
-    
-        command = f"lstchain_dl1_to_dl2 --input-files {input_fname} --path-models {rf_node} --output-dir {dir_dl2_scaled} --config {config_file}"
-        logger.info(command)
-    
-        subprocess.run(command, shell=True)
-    
-        dict_dchecks[run]["dl2_scaled"] = output_fname
-
-
+    if flag_scaled:
+        for ir, run in enumerate(dict_dchecks.keys()):
+        
+            input_fname  = dict_dchecks[run]["dl1b_scaled"]["runwise"]
+            output_fname = dir_dl2_scaled + input_fname.split("/")[-1].replace("dl1", "dl2", 1)
+            rf_node      = dict_dchecks[run]["simulations"]["rf"]
+        
+            # Check if the file exists and delete if exists (may be empty or half filled)
+            if os.path.exists(output_fname):
+                logger.info(f"File already exists, deleting and re-computing:\n-->{output_fname}")
+                os.remove(output_fname)
+        
+            logger.info(f"\nComputing dl2 for Run {run:5} (scaled data)")
+            logger.info(f"--> {output_fname}\n")
+        
+            command = f"lstchain_dl1_to_dl2 --input-files {input_fname} --path-models {rf_node} --output-dir {dir_dl2_scaled} --config {config_file}"
+            logger.info(command)
+        
+            subprocess.run(command, shell=True)
+        
+            dict_dchecks[run]["dl2_scaled"] = output_fname
 
 
-    for ir, run in enumerate(dict_dchecks.keys()):
-    
-        input_fname  = dict_dchecks[run]["dl1a"]["runwise"]
-        output_fname = dir_dl2 + input_fname.split("/")[-1].replace("dl1", "dl2", 1)
-        rf_node      = dict_dchecks[run]["simulations"]["rf"]
-    
-        # Check if the file exists and delete if exists (may be empty or half filled)
-        if os.path.exists(output_fname):
-            logger.info(f"File already exists, deleting and re-computing:\n-->{output_fname}")
-            os.remove(output_fname)
-    
-        logger.info(f"\nComputing dl2 for Run {run:5} (original data)")
-        logger.info(f"--> {output_fname}\n")
-    
-        command = f"lstchain_dl1_to_dl2 --input-files {input_fname} --path-models {rf_node} --output-dir {dir_dl2} --config {config_file}"
-        logger.info(command)
-    
-        subprocess.run(command, shell=True)
-    
-        dict_dchecks[run]["dl2"] = output_fname
+
+    if not flag_scaled:
+        for ir, run in enumerate(dict_dchecks.keys()):
+        
+            input_fname  = dict_dchecks[run]["dl1a"]["runwise"]
+            output_fname = dir_dl2 + input_fname.split("/")[-1].replace("dl1", "dl2", 1)
+            rf_node      = dict_dchecks[run]["simulations"]["rf"]
+        
+            # Check if the file exists and delete if exists (may be empty or half filled)
+            if os.path.exists(output_fname):
+                logger.info(f"File already exists, deleting and re-computing:\n-->{output_fname}")
+                os.remove(output_fname)
+        
+            logger.info(f"\nComputing dl2 for Run {run:5} (original data)")
+            logger.info(f"--> {output_fname}\n")
+        
+            command = f"lstchain_dl1_to_dl2 --input-files {input_fname} --path-models {rf_node} --output-dir {dir_dl2} --config {config_file}"
+            logger.info(command)
+        
+            subprocess.run(command, shell=True)
+        
+            dict_dchecks[run]["dl2"] = output_fname
 
 
 
@@ -302,31 +272,39 @@ def main(input_str):
     ra_str  = "{}".format(dict_source["ra"]).replace(" ", "")
     dec_str = "{}".format(dict_source["dec"]).replace(" ", "")
     
-    for ir, run in enumerate(dict_dchecks.keys()):
-    
-        # dir_run = dir_dl3 + f"{run:05}" + "/"    
-        dl2_fname        = dict_dchecks[run]["dl2"]
-        dl2_fname_scaled = dict_dchecks[run]["dl2_scaled"]
-    
-        output_dl3        = dir_dl3        + f"dl3_LST-1.Run{run:05}.fits"
-        output_dl3_scaled = dir_dl3_scaled + f"dl3_LST-1.Run{run:05}.fits"
+    if not flag_scaled:
+        for ir, run in enumerate(dict_dchecks.keys()):
         
-        logger.info(f"\nConverting dl2 for {run:5}")
-        logger.info(f"--> {output_dl3}\n--> {output_dl3_scaled}\n")
+            # dir_run = dir_dl3 + f"{run:05}" + "/"    
+            dl2_fname_scaled = dict_dchecks[run]["dl2_scaled"]
+            
+            output_dl3_scaled = dir_dl3_scaled + f"dl3_LST-1.Run{run:05}.fits"
+            
+            logger.info(f"--> {output_dl3}\n--> {output_dl3_scaled}\n")
         
-        command = f"lstchain_create_dl3_file --input-dl2 {dl2_fname} --input-irf-path {dir_irfs} --output-dl3-path {dir_dl3} "
-        command = command + f"--source-name {source_name} --source-ra {ra_str} --source-dec {dec_str} --config {config_file} --overwrite"
-        logger.info(command)
-        subprocess.run(command, shell=True)
+            command = f"lstchain_create_dl3_file --input-dl2 {dl2_fname_scaled} --input-irf-path {dir_irfs} --output-dl3-path {dir_dl3_scaled} "
+            command = command + f"--source-name {source_name} --source-ra {ra_str} --source-dec {dec_str} --config {config_file} --overwrite"
+            logger.info(command)
+            subprocess.run(command, shell=True)
     
-        command = f"lstchain_create_dl3_file --input-dl2 {dl2_fname_scaled} --input-irf-path {dir_irfs} --output-dl3-path {dir_dl3_scaled} "
-        command = command + f"--source-name {source_name} --source-ra {ra_str} --source-dec {dec_str} --config {config_file} --overwrite"
-        logger.info(command)
-        subprocess.run(command, shell=True)
-    
-        dict_dchecks[run]["dl3"]        = output_dl3
-        dict_dchecks[run]["dl3_scaled"] = output_dl3_scaled
-
+            dict_dchecks[run]["dl3_scaled"] = output_dl3_scaled
+            
+    if not flag_scaled:
+        for ir, run in enumerate(dict_dchecks.keys()):
+        
+            # dir_run = dir_dl3 + f"{run:05}" + "/"    
+            dl2_fname = dict_dchecks[run]["dl2"]
+        
+            output_dl3 = dir_dl3 + f"dl3_LST-1.Run{run:05}.fits"
+            
+            logger.info(f"\nConverting dl2 for {run:5}")
+            
+            command = f"lstchain_create_dl3_file --input-dl2 {dl2_fname} --input-irf-path {dir_irfs} --output-dl3-path {dir_dl3} "
+            command = command + f"--source-name {source_name} --source-ra {ra_str} --source-dec {dec_str} --config {config_file} --overwrite"
+            logger.info(command)
+            subprocess.run(command, shell=True)
+        
+            dict_dchecks[run]["dl3"] = output_dl3
